Shuffle headers to reduce dependencies.
