+++ include/vcpkg/base/files.h
@@ -9,7 +9,7 @@
 #include <vcpkg/base/file-contents.h>
 #include <vcpkg/base/lineinfo.h>
 #include <vcpkg/base/messages.h>
-#include <vcpkg/base/pragmas.h>
+#include <vcpkg/base/path.h>
 #include <vcpkg/base/stringview.h>
 
 #include <stdio.h>
@@ -21,12 +21,6 @@
 #include <system_error>
 #include <utility>
 #include <vector>
-
-#if defined(_WIN32)
-#define VCPKG_PREFERRED_SEPARATOR "\\"
-#else // ^^^ _WIN32 / !_WIN32 vvv
-#define VCPKG_PREFERRED_SEPARATOR "/"
-#endif // _WIN32
 
 namespace vcpkg
 {
@@ -43,61 +37,6 @@ namespace vcpkg
 
     private:
         std::error_code ec;
-    };
-
-    struct Path
-    {
-        Path();
-        Path(const Path&);
-        Path(Path&&);
-        Path& operator=(const Path&);
-        Path& operator=(Path&&);
-
-        Path(const StringView sv);
-        Path(const std::string& s);
-        Path(std::string&& s);
-        Path(const char* s);
-        Path(const char* first, size_t size);
-
-        const std::string& native() const& noexcept;
-        std::string&& native() && noexcept;
-        operator StringView() const noexcept;
-
-        const char* c_str() const noexcept;
-
-        std::string generic_u8string() const;
-
-        bool empty() const noexcept;
-
-        Path operator/(StringView sv) const&;
-        Path operator/(StringView sv) &&;
-        Path operator+(StringView sv) const&;
-        Path operator+(StringView sv) &&;
-
-        Path& operator/=(StringView sv);
-        Path& operator+=(StringView sv);
-
-        void replace_filename(StringView sv);
-        void remove_filename();
-        void make_preferred();
-        void clear();
-        Path lexically_normal() const;
-
-        // Sets *this to parent_path, returns whether anything was removed
-        bool make_parent_path();
-
-        StringView parent_path() const;
-        StringView filename() const;
-        StringView extension() const;
-        StringView stem() const;
-
-        bool is_absolute() const;
-        bool is_relative() const;
-
-        friend const char* to_printf_arg(const Path& p) noexcept;
-
-    private:
-        std::string m_str;
     };
 
     bool is_symlink(FileType s);
