+++ include/vcpkg/base/system.process.h
@@ -3,7 +3,7 @@
 #include <vcpkg/base/fwd/system.process.h>
 
 #include <vcpkg/base/expected.h>
-#include <vcpkg/base/files.h>
+#include <vcpkg/base/path.h>
 #include <vcpkg/base/span.h>
 #include <vcpkg/base/stringview.h>
