+++ include/vcpkg/base/unicode.h
@@ -29,8 +29,6 @@ namespace vcpkg::Unicode
         UnexpectedEof = 6,
     };
 
-    const std::error_category& utf8_category() noexcept;
-
     Utf8CodeUnitKind utf8_code_unit_kind(unsigned char code_unit) noexcept;
 
     constexpr int utf8_code_unit_count(Utf8CodeUnitKind kind) noexcept { return static_cast<int>(kind); }
@@ -103,11 +101,6 @@ namespace vcpkg::Unicode
     }
 
     char32_t utf16_surrogates_to_code_point(char32_t leading, char32_t trailing);
-
-    inline std::error_code make_error_code(utf8_errc err) noexcept
-    {
-        return std::error_code(static_cast<int>(err), utf8_category());
-    }
 
     /*
         There are two ways to parse utf-8: we could allow unpaired surrogates (as in [wtf-8]) -- this is important
@@ -188,12 +181,4 @@ namespace vcpkg::Unicode
     constexpr bool operator==(Utf8Decoder::sentinel s, const Utf8Decoder& d) { return d == s; }
     constexpr bool operator!=(const Utf8Decoder& d, Utf8Decoder::sentinel) { return !d.is_eof(); }
     constexpr bool operator!=(Utf8Decoder::sentinel s, const Utf8Decoder& d) { return d != s; }
-}
-
-namespace std
-{
-    template<>
-    struct is_error_code_enum<vcpkg::Unicode::utf8_errc> : std::true_type
-    {
-    };
 }
