+++ include/vcpkg/binarycaching.h
@@ -8,7 +8,7 @@
 
 #include <vcpkg/base/downloads.h>
 #include <vcpkg/base/expected.h>
-#include <vcpkg/base/files.h>
+#include <vcpkg/base/path.h>
 
 #include <vcpkg/packagespec.h>
