+++ include/vcpkg/configuration.h
@@ -5,14 +5,23 @@
 #include <vcpkg/fwd/configuration.h>
 #include <vcpkg/fwd/vcpkgcmdarguments.h>
 
-#include <vcpkg/base/files.h>
 #include <vcpkg/base/json.h>
 #include <vcpkg/base/optional.h>
+#include <vcpkg/base/path.h>
 
 #include <vcpkg/registries.h>
 
 namespace vcpkg
 {
+    /// <summary>
+    /// A registry package pattern (e.g.: boost*) and the in-file location where it was declared.
+    /// </summary>
+    struct PackagePatternDeclaration
+    {
+        std::string pattern;
+        std::string location;
+    };
+
     struct RegistryConfig
     {
         // Missing kind means "null"
@@ -24,7 +33,7 @@ namespace vcpkg
         Optional<std::string> reference;
         Optional<std::string> repo;
         Optional<std::vector<std::string>> packages;
-        Optional<std::vector<Json::PackagePatternDeclaration>> package_declarations;
+        Optional<std::vector<PackagePatternDeclaration>> package_declarations;
 
         Json::Value serialize() const;
 
@@ -87,4 +96,7 @@ namespace vcpkg
                                                 vcpkg::MessageSink& messageSink);
 
     std::vector<std::string> find_unknown_fields(const Configuration& config);
+
+    // Exposed for testing
+    bool is_package_pattern(StringView sv);
 }
