+++ src/vcpkg/commands.hash.cpp
@@ -1,3 +1,4 @@
+#include <vcpkg/base/files.h>
 #include <vcpkg/base/hash.h>
 #include <vcpkg/base/util.h>
