+++ src/vcpkg/commands.portsdiff.cpp
@@ -1,4 +1,4 @@
-#include <vcpkg/base/sortedvector.h>
+#include <vcpkg/base/files.h>
 #include <vcpkg/base/system.process.h>
 #include <vcpkg/base/util.h>
