+++ src/vcpkg/packagespec.cpp
@@ -26,6 +26,35 @@ namespace vcpkg
         return fmt::format("{}[{}]", package_name, feature_name);
     }
 
+    static InternalFeatureSet normalize_feature_list(View<std::string> fs, ImplicitDefault id)
+    {
+        InternalFeatureSet ret;
+        bool core = false;
+        for (auto&& f : fs)
+        {
+            if (f == "core")
+            {
+                core = true;
+            }
+            ret.emplace_back(f);
+        }
+
+        if (!core)
+        {
+            ret.emplace_back("core");
+            if (id == ImplicitDefault::YES)
+            {
+                ret.emplace_back("default");
+            }
+        }
+        return ret;
+    }
+
+    FullPackageSpec::FullPackageSpec(PackageSpec spec, View<std::string> features, ImplicitDefault id)
+        : package_spec(std::move(spec)), features(normalize_feature_list(features, id))
+    {
+    }
+
     void FullPackageSpec::expand_fspecs_to(std::vector<FeatureSpec>& out) const
     {
         for (auto&& feature : features)
@@ -51,30 +80,6 @@ namespace vcpkg
         return left.name() == right.name() && left.triplet() == right.triplet();
     }
 
-    static InternalFeatureSet normalize_feature_list(View<std::string> fs, ImplicitDefault id)
-    {
-        InternalFeatureSet ret;
-        bool core = false;
-        for (auto&& f : fs)
-        {
-            if (f == "core")
-            {
-                core = true;
-            }
-            ret.emplace_back(f);
-        }
-
-        if (!core)
-        {
-            ret.emplace_back("core");
-            if (id == ImplicitDefault::YES)
-            {
-                ret.emplace_back("default");
-            }
-        }
-        return ret;
-    }
-
     ExpectedL<FullPackageSpec> ParsedQualifiedSpecifier::to_full_spec(Triplet default_triplet, ImplicitDefault id) const
     {
         if (platform)
@@ -84,7 +89,7 @@ namespace vcpkg
 
         const Triplet t = triplet ? Triplet::from_canonical_name(*triplet.get()) : default_triplet;
         const View<std::string> fs = !features.get() ? View<std::string>{} : *features.get();
-        return FullPackageSpec{{name, t}, normalize_feature_list(fs, id)};
+        return FullPackageSpec{{name, t}, fs, id};
     }
 
     ExpectedL<PackageSpec> ParsedQualifiedSpecifier::to_package_spec(Triplet default_triplet) const
@@ -256,52 +261,4 @@ namespace vcpkg
         parser.skip_tabs_spaces();
         return ret;
     }
-
-    bool operator==(const DependencyConstraint& lhs, const DependencyConstraint& rhs)
-    {
-        if (lhs.type != rhs.type) return false;
-        if (lhs.value != rhs.value) return false;
-        return lhs.port_version == rhs.port_version;
-    }
-
-    Optional<Version> DependencyConstraint::try_get_minimum_version() const
-    {
-        if (type == VersionConstraintKind::None)
-        {
-            return nullopt;
-        }
-
-        return Version{
-            value,
-            port_version,
-        };
-    }
-
-    FullPackageSpec Dependency::to_full_spec(Triplet target, Triplet host_triplet, ImplicitDefault id) const
-    {
-        return FullPackageSpec{{name, host ? host_triplet : target}, normalize_feature_list(features, id)};
-    }
-
-    bool operator==(const Dependency& lhs, const Dependency& rhs)
-    {
-        if (lhs.name != rhs.name) return false;
-        if (lhs.features != rhs.features) return false;
-        if (!structurally_equal(lhs.platform, rhs.platform)) return false;
-        if (lhs.extra_info != rhs.extra_info) return false;
-        if (lhs.constraint != rhs.constraint) return false;
-        if (lhs.host != rhs.host) return false;
-
-        return true;
-    }
-    bool operator!=(const Dependency& lhs, const Dependency& rhs);
-
-    bool operator==(const DependencyOverride& lhs, const DependencyOverride& rhs)
-    {
-        if (lhs.version_scheme != rhs.version_scheme) return false;
-        if (lhs.port_version != rhs.port_version) return false;
-        if (lhs.name != rhs.name) return false;
-        if (lhs.version != rhs.version) return false;
-        return lhs.extra_info == rhs.extra_info;
-    }
-    bool operator!=(const DependencyOverride& lhs, const DependencyOverride& rhs);
 }
