+++ src/vcpkg/portfileprovider.cpp
@@ -1,4 +1,4 @@
-#include <vcpkg/base/json.h>
+#include <vcpkg/base/files.h>
 #include <vcpkg/base/messages.h>
 #include <vcpkg/base/system.debug.h>
